+++ packages/mobile/buttons/__stories__/ButtonGroup.stories.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react-native-a11y/has-accessibility-hint */
 import React from 'react';
 import { GestureResponderEvent } from 'react-native';