+++ packages/mobile/controls/__stories__/RadioGroup.stories.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react-native-a11y/has-accessibility-hint */
 import React, { useState } from 'react';
 import { useToggler } from '@cbhq/cds-common/hooks/useToggler';